:truck: moved getters into separated forlders
